+++ config/routes.rb
@@ -1,7 +1,5 @@
 Rails.application.routes.draw do
-  unless respond_to?(:has_named_route?) && has_named_route?("ahoy_email_engine")
-    mount AhoyEmail::Engine => "/ahoy"
-  end
+  mount AhoyEmail::Engine => "/ahoy" if AhoyEmail.api
 end
 
 AhoyEmail::Engine.routes.draw do