+++ lib/generators/ahoy_email/install_generator.rb
@@ -22,7 +22,13 @@ module AhoyEmail
       end
 
       def copy_migration
-        migration_template "install.rb", "db/migrate/create_ahoy_messages.rb"
+        migration_template "install.rb", "db/migrate/create_ahoy_messages.rb", migration_version: migration_version
+      end
+
+      def migration_version
+        if ActiveRecord::VERSION::MAJOR >= 5
+          "[#{ActiveRecord::VERSION::MAJOR}.#{ActiveRecord::VERSION::MINOR}]"
+        end
       end
     end
   end