+++ test/click_test.rb
@@ -1,0 +1,28 @@
+require_relative "test_helper"
+
+class ClickTest < Minitest::Test
+  def test_default
+    message = ClickMailer.welcome.deliver_now
+    refute_body "click", message
+  end
+
+  def test_basic
+    message = ClickMailer.basic.deliver_now
+    assert_body "click", message
+  end
+
+  def test_mailto
+    message = ClickMailer.mailto.deliver_now
+    assert_body '<a href="mailto:hi@example.org">', message
+  end
+
+  def test_app
+    message = ClickMailer.app.deliver_now
+    assert_body '<a href="fb://profile/33138223345">', message
+  end
+
+  def test_schemeless
+    message = ClickMailer.schemeless.deliver_now
+    assert_body "click", message
+  end
+end