+++ test/internal/app/mailers/user_mailer.rb
@@ -1,0 +1,21 @@
+class UserMailer < ApplicationMailer
+  track user: -> { @some_user }, only: [:dynamic]
+
+  def welcome
+    mail
+  end
+
+  def user_var(user)
+    @user = user
+    mail
+  end
+
+  def to_field(to)
+    mail to: to
+  end
+
+  def dynamic(some_user)
+    @some_user = some_user
+    mail
+  end
+end