+++ test/internal/db/schema.rb
@@ -1,30 +1,24 @@
 ActiveRecord::Schema.define do
   create_table :ahoy_messages, force: true do |t|
-    t.string :token
-
-    # user
+    t.references :user, polymorphic: true
     t.text :to
-    t.integer :user_id
-    t.string :user_type
-
-    # optional - feel free to remove
     t.string :mailer
     t.text :subject
-    t.text :content
+    t.timestamp :sent_at
 
-    # optional
+    # extra
+    t.integer :coupon_id
+
+    # legacy
+    t.text :content
     t.string :utm_source
     t.string :utm_medium
     t.string :utm_term
     t.string :utm_content
     t.string :utm_campaign
-
-    # timestamps
-    t.timestamp :sent_at
-    t.timestamp :opened_at
-    t.timestamp :clicked_at
   end
 
-  add_index :ahoy_messages, [:token]
-  add_index :ahoy_messages, [:user_id, :user_type]
+  create_table :users, force: true do |t|
+    t.string :email
+  end
 end