+++ test/test_helper.rb
@@ -1,14 +1,48 @@
 require "bundler/setup"
+require "combustion"
 Bundler.require(:default)
 require "minitest/autorun"
 require "minitest/pride"
-require "combustion"
-
-Minitest::Test = Minitest::Unit::TestCase unless defined?(Minitest::Test)
 
 Combustion.path = "test/internal"
-Combustion.initialize! :all
+Combustion.initialize! :all do
+  if config.active_record.sqlite3.respond_to?(:represent_boolean_as_integer)
+    config.active_record.sqlite3.represent_boolean_as_integer = true
+  end
+end
+
+require_relative "support/mongoid" if defined?(Mongoid)
 
 ActionMailer::Base.delivery_method = :test
 
-ActiveRecord::Base.belongs_to_required_by_default = true if ActiveRecord::VERSION::MAJOR >= 5
+class Minitest::Test
+  def setup
+    Ahoy::Message.delete_all
+  end
+
+  def ahoy_message
+    Ahoy::Message.last
+  end
+
+  def refute_body(str, message)
+    refute_match str, message.body.decoded
+  end
+
+  def assert_body(str, message)
+    assert_match str, message.body.decoded
+  end
+
+  def params_supported?
+    Rails.version > "5.1.0"
+  end
+
+  def with_default(options)
+    previous_options = AhoyEmail.default_options.dup
+    begin
+      AhoyEmail.default_options.merge!(options)
+      yield
+    ensure
+      AhoyEmail.default_options = previous_options
+    end
+  end
+end